more verbose script output and error hanlding

+++ alpine-xtables-geoip-update.sh
@@ -14,16 +14,34 @@ export PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 URL=https://mailfud.org/geoip-legacy/GeoIP-legacy.csv.gz
 DBDIR=/usr/share/GeoIP
 
+echo "INFO: ensure packages are installed"
 apk add perl-net-cidr-lite xtables-addons perl perl-doc perl-text-csv_xs unzip
 
 test -d $DBDIR || mkdir -p $DBDIR
 cd $DBDIR || exit 1
 rm -f GeoIP-legacy.*
+
 wget -q $URL
+if [ "$?" -ne 0 ]
+then
+   >&2 echo "ERROR: could not download $URL"
+   exit 1
+else
+   echo "INFO: downloaded $URL"
+fi
+
 gunzip GeoIP-legacy.csv.gz
 cat $DBDIR/GeoIP-legacy.csv | cut -d, -f1,2,5 > $DBDIR/dbip-country-lite.csv
 rm -f $DBDIR/GeoIP-legacy.*
-/usr/libexec/xtables-addons/xt_geoip_build -D /usr/share/xt_geoip $DBDIR/GeoIP-legacy.csv
-rm -f *.csv *.txt *.gz
 
+/usr/libexec/xtables-addons/xt_geoip_build -D /usr/share/xt_geoip $DBDIR/GeoIP-legacy.csv | grep -q 'ranges for ZW'
+if [ "$?" -ne 0 ]
+then
+   >&2 echo "ERROR: failed to build GeoIP database files in /usr/share/xt_geoip"
+   exit 1
+else
+   echo "INFO: Finished building xt_geoip database files"
+   cd $DBDIR
+   rm -f *.csv *.txt *.gz
+fi
